AnaliticsFragment.kt improvements

+++ app/src/main/java/com/invest/advisor/data/network/yahooResponse/FreeCashflow.kt
@@ -4,7 +4,7 @@ package com.invest.advisor.data.network.yahooResponse
 import com.google.gson.annotations.SerializedName
 
 data class FreeCashflow(
-    val raw: Long,
+    val raw: Double,
     val fmt: String,
     val longFmt: String
 )
